...Topbar and side navbar added and also folder structure has been has been modified for request details ticket page
