+++ package.json
@@ -28,7 +28,9 @@
     "react": "^19.1.1",
     "react-bootstrap": "^2.10.10",
     "react-dom": "^19.1.1",
+    "react-icons": "^5.5.0",
     "react-password-strength-bar": "^0.4.1",
+    "react-pro-sidebar": "^0.7.1",
     "react-quill-new": "^3.6.0",
     "react-redux": "^9.2.0",
     "react-router-dom": "^6.30.1",
