--- src/Components/HomePage/Navbar.js
@@ -1,318 +1,0 @@
-import React, { useEffect, useRef, useState } from "react";
-import {
-  AppBar,
-  Toolbar,
-  IconButton,
-  Tooltip,
-  TextField,
-  Box,
-  Menu,
-  MenuItem,
-  Drawer,
-  List,
-  ListItem,
-  ListItemButton,
-  ListItemText,
-  Slide
-} from "@mui/material";
-import MenuIcon from "@mui/icons-material/Menu";
-import SearchIcon from "@mui/icons-material/Search";
-import NotificationsIcon from "@mui/icons-material/Notifications";
-import SettingsIcon from "@mui/icons-material/Settings";
-import AccountCircleIcon from "@mui/icons-material/AccountCircle";
-import HomeFilledIcon from "@mui/icons-material/HomeFilled";
-import ConfirmationNumberIcon from "@mui/icons-material/ConfirmationNumber";
-import EmojiObjectsIcon from "@mui/icons-material/EmojiObjects";
-import AttachMoneyIcon from "@mui/icons-material/AttachMoney";
-import FlagIcon from "@mui/icons-material/Flag";
-import Divider from '@mui/material/Divider';
-import InputBase from '@mui/material/InputBase';
-import { alpha,styled } from '@mui/material/styles';
-import { NavLink, useLocation, useNavigate } from "react-router-dom";
-import { useDispatch } from "react-redux";
-import * as actions from "../../actions"
-
-
- const Search = styled('div')(({ theme }) => ({
-  position: 'relative',
-  borderRadius: theme.shape.borderRadius,
-  backgroundColor: alpha(theme.palette.common.white, 0.15),
-  '&:hover': {
-    backgroundColor: alpha(theme.palette.common.white, 0.25),
-  },
-  marginLeft: 0,
-  width: '100%',
-  [theme.breakpoints.up('sm')]: {
-    marginLeft: theme.spacing(1),
-    width: 'auto',
-  },
-}));
-
- const StyledInputBase = styled(InputBase)(({ theme }) => ({
-  color: 'inherit',
-  width: '100%',
-  '& .MuiInputBase-input': {
-    padding: theme.spacing(1, 1, 1, 0),
-    // vertical padding + font size from searchIcon
-    paddingLeft: `calc(1em + ${theme.spacing(4)})`,
-    transition: theme.transitions.create('width'),
-    [theme.breakpoints.up('sm')]: {
-      width: '12ch',
-      '&:focus': {
-        width: '20ch',
-      },
-    },
-  },
-}));
-
- const SearchIconWrapper = styled('div')(({ theme }) => ({
-  padding: theme.spacing(0, 2),
-  height: '100%',
-  position: 'absolute',
-  pointerEvents: 'none',
-  display: 'flex',
-  alignItems: 'center',
-  justifyContent: 'center',
-}));
-
-export default function Navbar() {
-  const navigate = useNavigate();
-  const location = useLocation();
-  const dispatch=useDispatch();
-
-  const [showInput, setShowInput] = useState(false);
-  const [anchorEl, setAnchorEl] = useState(null);
-  const [pendingRoute, setPendingRoute] = useState(null);
-  const [drawerOpen, setDrawerOpen] = useState(false);
-
-  const [highlightStyle, setHighlightStyle] = useState({});
-  const navRefs = useRef({});
-
-  // Nav items with proper icons
-
-
-  const navItems = [
-    { text: "home", icon: <HomeFilledIcon /> },
-    { text: "request", icon: <ConfirmationNumberIcon /> },
-    { text: "solution", icon: <EmojiObjectsIcon /> },
-    { text: "asset", icon: <AttachMoneyIcon /> },
-    { text: "report", icon: <FlagIcon /> },
-  ];
-
-  const handleMenu = (event) => setAnchorEl(event.currentTarget);
-  const handleClose = () => setAnchorEl(null);
-  const handleNavigate = (path) => {
-    setPendingRoute(path);
-    handleClose();
-  };
-
-  // navigate after closing
-  useEffect(() => {
-    if (!anchorEl && pendingRoute) {
-      navigate(pendingRoute);
-      setPendingRoute(null);
-    }
-  }, [anchorEl, pendingRoute, navigate]);
-
-  // highlight effect
-    useEffect(() => {
-      const updateHighlight = () => {
-        const activePath = location.pathname;
-        const activeItem = navItems.find(
-          (item) => `/${item.text}` === activePath
-        );
-        if (!activeItem) return;
-
-        const el = navRefs.current[activeItem.text];
-        if (el) {
-          setHighlightStyle({
-            left: el.offsetLeft,
-            width: el.offsetWidth,
-          });
-        }
-      };
-
-      const id = requestAnimationFrame(updateHighlight);
-      return () => cancelAnimationFrame(id);
-    }, [location.pathname, navItems]);
-
-
-  return (
-    <>
-      <AppBar position="static" sx={{ backgroundColor: "#D95D2E" }}>
-        <Toolbar sx={{ display: "flex", justifyContent: "space-between" }}>
-          {/* Hamburger for mobile */}
-          <Box sx={{ display: { xs: "block", md: "none" } }}>
-            <IconButton
-              onClick={() => setDrawerOpen(true)}
-              sx={{ color: "white" }}
-            >
-              <MenuIcon />
-            </IconButton>
-          </Box>
-
-          {/* Desktop Nav */}
-          <Box
-            sx={{
-              position: "relative",
-              display: { xs: "none", md: "flex" },
-              gap: 4,
-            }}
-          >
-            <div
-              style={{
-                position: "absolute",
-                top: 0,
-                bottom: 0,
-                background: "white",
-                borderRadius: "10px",
-                padding: "5px",
-                height: "27px",
-                transition: "all 0.3s ease",
-                ...highlightStyle,
-              }}
-            />
-            {navItems.map((item) => {
-              const path = `/${item.text}`;
-              const isActive = location.pathname.startsWith(path);
-              return (
-                <NavLink
-                  key={item.text}
-                  to={path}
-                  ref={(el) => (navRefs.current[item.text] = el)}
-                  style={{
-                    position: "relative",
-                    padding: "6px 14px",
-                    borderRadius: "6px",
-                    textAlign: "center",
-                    color: isActive ? "#D95D2E" : "rgba(255,255,255,0.8)",
-                    fontWeight: isActive ? "bold" : "normal",
-                    textDecoration: "none",
-                    zIndex: 1,
-                  }}
-                >
-                  {item.text.charAt(0).toUpperCase() + item.text.slice(1)}
-                </NavLink>
-              );
-            })}
-          </Box>
-
-          {/* Right side icons */}
-          <Box sx={{ display: "flex", alignItems: "center", gap: 1 }}>
-            {/* <Box
-              sx={{
-                width: showInput ? 200 : 0,
-                opacity: showInput ? 1 : 0,
-                overflow: "hidden",
-                transition: "width 0.3s ease, opacity 0.3s ease",
-              }}
-            >
-            </Box> */}
-            <Toolbar style={{paddingRight:'0px'}}>
-              <Search>
-                <SearchIconWrapper>
-                  <SearchIcon />
-                </SearchIconWrapper>
-                <StyledInputBase
-                  id='searchfield'
-                  placeholder="Search…"
-                  onChange={(e)=>e.preventDefault()}
-                  inputProps={{ 'aria-label': 'search' }}
-                />
-              </Search>
-            </Toolbar>
-
-            {/* <Tooltip title="Search">
-              <IconButton
-                onClick={() => setShowInput((prev) => !prev)}
-                sx={{ color: "white" }}
-              >
-                <SearchIcon />
-              </IconButton>
-            </Tooltip> */}
-
-            <Tooltip title="Notifications">
-              <IconButton sx={{ color: "white" }}>
-                <NotificationsIcon />
-              </IconButton>
-            </Tooltip>
-
-            <Tooltip title="Settings">
-              <IconButton sx={{ color: "white" }}>
-                <SettingsIcon />
-              </IconButton>
-            </Tooltip>
-
-            <Tooltip title="Account">
-              <IconButton onClick={handleMenu} sx={{ color: "white" }}>
-                <AccountCircleIcon />
-              </IconButton>
-            </Tooltip>
-            <Menu
-              anchorEl={anchorEl}
-              open={Boolean(anchorEl)}
-              onClose={handleClose}
-              anchorOrigin={{ vertical: "top", horizontal: "right" }}
-              transformOrigin={{ vertical: "top", horizontal: "right" }}
-            >
-              <MenuItem onClick={() => handleNavigate("/account-details")}>
-                My account
-              </MenuItem>
-              <MenuItem onClick={() =>
-                { 
-                 dispatch(actions.logout())
-                 navigate("/login", { replace: true });
-                //  handleNavigate("/login")
-                }
-                 }>Log Out</MenuItem>
-            </Menu>
-          </Box>
-        </Toolbar>
-      </AppBar>
-
-      {/* Slide Drawer for mobile */}
-        <Drawer
-          anchor="right"
-          open={drawerOpen}
-          onClose={() => setDrawerOpen(false)}
-        >
-          <Box sx={{ width: 250 }}>
-            <List>
-         {navItems.map((item, index) => (
-              <React.Fragment key={item.text}>
-                <ListItem disablePadding>
-                  <ListItemButton
-                    onClick={() => {
-                      navigate(`/${item.text}`);
-                      setDrawerOpen(false);
-                    }}
-                  >
-                    <div
-                      style={{
-                        display: "flex",
-                        justifyContent: "center",
-                        alignItems: "center",
-                        gap: "8px"
-                      }}
-                    >
-                      <span>{item.icon}</span>
-                      <ListItemText
-                        primary={
-                          item.text.charAt(0).toUpperCase() + item.text.slice(1)
-                        }
-                      />
-                    </div>
-                  </ListItemButton>
-                </ListItem>
-
-                {/* Divider should be outside ListItem */}
-                {index < navItems.length - 1 && <Divider />}
-              </React.Fragment>
-            ))}
-
-            </List>
-          </Box>
-        </Drawer>
-    </>
-  );
-}
