+++ src/Components/HomePage/sidebar.css
@@ -1,0 +1,20 @@
+.pro-sidebar .pro-sidebar-inner {
+  background-image: url('/public/images/orange1.jpg') !important;
+  background-size: cover !important;
+  background-repeat: no-repeat !important;
+  background-position: center !important;
+}
+
+.pro-sidebar-inner .pro-sidebar-layout .active {
+    background-image: linear-gradient(0deg, #eed295 0%, #f0ead4 100%);
+    /* transition: 0.3s ease-in-out; */
+    border-radius: 6px;
+    color: black !important;
+  }
+  .pro-sidebar .ps-menu-button {
+  transition: background-color 0.3s ease-in-out, color 0.3s ease-in-out;
+}
+
+.pro-sidebar .ps-menu-button:hover {
+  background-color: #f5e5b8; /* optional hover effect */
+}
