+++ src/Components/Requests/RequestPage.js
@@ -18,6 +18,7 @@ const ContentWrapper = styled.div`
   }
 `;
 
+
 export default function Request() {
   return (
     <Paper
@@ -26,6 +27,7 @@ export default function Request() {
         p: 3,
         m: 1,
         borderRadius: 3,
+        
       }}
     >
       <ContentWrapper>
