+++ src/Components/Solution.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import Navbar from './HomePage/Navbar'
+// import Navbar from './HomePage/SideNavbar'
 import pdfFile from "../assets/mrt-pdf-example.pdf"
 
 export default function Solution() {
   return (
    <div>
-         <Navbar/>
+         {/* <Navbar/> */}
          {/* <h4>Solution Page</h4> */}
          <iframe src={pdfFile} width="100%" height="500px" title="PDF Viewer" />
     </div>
