+++ src/Config/SuccessFailureSnackbar.js
@@ -21,9 +21,9 @@ export default function SuccessFailureSnackbar() {
 
   return (
     <>
-    {/* <Button onClick={()=>setOpen(true)}>Click</Button> */}
     <Snackbar 
         open={opener} 
+        
         autoHideDuration={3000} 
         onClose={()=>{
           dispatch(actions.closeSnackbar())
@@ -35,7 +35,7 @@ export default function SuccessFailureSnackbar() {
           dispatch(actions.closeSnackbar())
         }}
           severity={status}
-          variant="filled"
+          // variant='outlined'
           sx={{ width: '100%' }}
         >
           {message}
