+++ src/Config/domain.js
@@ -1,3 +1,3 @@
 // domain name + port no 
 // export const domainurl = 'https://jsonplaceholder.typicode.com';
-export const domainurl = 'http://172.31.208.1:3001';
+export const domainurl = 'http://172.31.208.1:3000';
