+++ src/styled_components/requesttable.styled.js
@@ -23,6 +23,7 @@ export const StyledTableContainer = styled(TableContainer)`
   // border: 1px solid #ddd;
   border-radius: 6px;
   overflow: hidden;
+  overflow-x: auto;
   width: 100%;
   min-height: 470px;   
 `;
